LPS-54752 Move *ComponentProvider to util package.
I chose this name because current Util classes in portal serve the same purpose as ComponentProviders
