+++ modules/apps/bookmarks/bookmarks-web/docroot/html/portlet/init.jsp
@@ -42,7 +42,7 @@ page import="com.liferay.bookmarks.service.permission.BookmarksEntryPermissionCh
 page import="com.liferay.bookmarks.service.permission.BookmarksFolderPermissionChecker" %><%@
 page import="com.liferay.bookmarks.settings.BookmarksGroupServiceSettings" %><%@
 page import="com.liferay.bookmarks.web.portlet.util.BookmarksUtil" %><%@
-page import="com.liferay.bookmarks.web.settings.BookmarksWebComponentProvider" %><%@
+page import="com.liferay.bookmarks.web.util.BookmarksWebComponentProvider" %><%@
 page import="com.liferay.portal.kernel.bean.BeanParamUtil" %><%@
 page import="com.liferay.portal.kernel.dao.search.ResultRow" %><%@
 page import="com.liferay.portal.kernel.dao.search.SearchEntry" %><%@
