+++ modules/apps/wiki/wiki-test/test/unit/com/liferay/wiki/engine/creole/util/test/CreoleTestUtil.java
@@ -18,7 +18,7 @@ import com.liferay.wiki.configuration.WikiGroupServiceConfiguration;
 import com.liferay.wiki.parser.creole.ast.WikiPageNode;
 import com.liferay.wiki.parser.creole.parser.Creole10Lexer;
 import com.liferay.wiki.parser.creole.parser.Creole10Parser;
-import com.liferay.wiki.service.provider.WikiServiceComponentProvider;
+import com.liferay.wiki.service.util.WikiServiceComponentProvider;
 
 import java.io.IOException;
 import java.io.InputStream;
