+++ modules/apps/wiki/wiki-web/docroot/html/portlet/wiki_display/init.jsp
@@ -22,7 +22,7 @@ page import="com.liferay.wiki.model.WikiNode" %><%@
 page import="com.liferay.wiki.model.WikiPage" %><%@
 page import="com.liferay.wiki.service.WikiNodeServiceUtil" %><%@
 page import="com.liferay.wiki.service.WikiPageLocalServiceUtil" %><%@
-page import="com.liferay.wiki.web.provider.WikiWebComponentProvider" %>
+page import="com.liferay.wiki.web.util.WikiWebComponentProvider" %>
 
 <%
 WikiWebComponentProvider wikiWebComponentProvider = WikiWebComponentProvider.getWikiWebComponentProvider();
